Software repository - delete version: confirmation popup added
Bug: DCS-1009

+++ ui/webapp/src/app/modules/repositories/components/software-list/software-list.component.ts
@@ -1,6 +1,8 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { finalize } from 'rxjs/operators';
 import { AuthenticationService } from 'src/app/modules/core/services/authentication.service';
+import { MessageBoxMode } from 'src/app/modules/shared/components/messagebox/messagebox.component';
+import { MessageboxService } from 'src/app/modules/shared/services/messagebox.service';
 import { ManifestKey } from '../../../../models/gen.dtos';
 import { LoggingService } from '../../../core/services/logging.service';
 import { DownloadService } from '../../../shared/services/download.service';
@@ -37,18 +39,27 @@ export class SoftwareListComponent implements OnInit {
 
   public exporting: ManifestKey = null;
 
-  constructor(private softwareService: SoftwareService,
-     private loggingService: LoggingService,
-     private downloadService: DownloadService,
-     private authService: AuthenticationService) { }
+  constructor(
+    private messageBoxService: MessageboxService,
+    private softwareService: SoftwareService,
+    private loggingService: LoggingService,
+    private downloadService: DownloadService,
+    private authService: AuthenticationService) { }
 
   ngOnInit() {
   }
 
   delete(softwareVersion: ManifestKey): void {
-    this.softwareService.deleteSoftwareVersion(this.softwareRepositoryName, softwareVersion).subscribe(r => {
-      this.log.message('Successfully deleted ' + softwareVersion.name + ':' + softwareVersion.tag);
-      this.deleted.emit();
+    this.messageBoxService.open({
+      title: 'Delete',
+      message: 'Do you really want to delete the software version ' + softwareVersion.name + ':' + softwareVersion.tag + '?',
+      mode: MessageBoxMode.CONFIRM,
+    }).subscribe(r => {
+      if (r) {
+        this.softwareService.deleteSoftwareVersion(this.softwareRepositoryName, softwareVersion).subscribe(r => {
+          this.deleted.emit();
+        });
+      }
     });
   }
